feat: map backspace key to backspace function
The backspace key now operates the backspace function.
This has resulted in a couple of minor changes. Most notably,
the ANS function is now represented with "A", so that the
letters can't be separated, thus causing incorrect syntax.
The other notable change is evaluateRPN now returns a string of
the evaluated value rather than a float, so that evaluated can
be backspaced without crashing the page.