feat: bind A key to Ans function
Bind A key to Ans function.
Also fixed the issue of Ans being labelled A by treating it
as a special case of button event.
Also fixed a bug where event.target.value was incorrectly
referred to as "event.type.value". I have no idea how long that
was there for.